Image painter

+++ src/flat/math/stroke/Stroker.java
@@ -996,12 +996,12 @@ public final class Stroker implements PathConsumer {
         // if these vectors are too small, normalize them, to avoid future
         // precision problems.
         if (Math.abs(dxs) < 0.1f && Math.abs(dys) < 0.1f) {
-            double len = (double)Math.sqrt(dxs*dxs + dys*dys);
+            double len = Math.sqrt(dxs*dxs + dys*dys);
             dxs /= len;
             dys /= len;
         }
         if (Math.abs(dxf) < 0.1f && Math.abs(dyf) < 0.1f) {
-            double len = (double)Math.sqrt(dxf*dxf + dyf*dyf);
+            double len = Math.sqrt(dxf*dxf + dyf*dyf);
             dxf /= len;
             dyf /= len;
         }
@@ -1077,12 +1077,12 @@ public final class Stroker implements PathConsumer {
         // if these vectors are too small, normalize them, to avoid future
         // precision problems.
         if (Math.abs(dxs) < 0.1f && Math.abs(dys) < 0.1f) {
-            double len = (double)Math.sqrt(dxs*dxs + dys*dys);
+            double len = Math.sqrt(dxs*dxs + dys*dys);
             dxs /= len;
             dys /= len;
         }
         if (Math.abs(dxf) < 0.1f && Math.abs(dyf) < 0.1f) {
-            double len = (double)Math.sqrt(dxf*dxf + dyf*dyf);
+            double len = Math.sqrt(dxf*dxf + dyf*dyf);
             dxf /= len;
             dyf /= len;
         }
